services/ddi-infer/requirements.txt
# Inference service dependencies (GPU)
# Torch & torchvision come from base image (pytorch/pytorch:2.3.0-cuda11.8-cudnn8-runtime)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers
tokenizers
pillow==10.2.0
pandas==2.1.2
scikit-learn==1.3.2
rdkit==2023.9.5
torch-geometric==2.4.0
pyg-lib==0.3.1
